pyproject.toml
[project]
name = "agentic-interview"
version = "0.1.0"
description = "AI-powered interview platform"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "asgiref>=3.8.1",
    "channels>=4.2.2",
    "channels-redis>=4.2.1",
    "cryptography>=45.0.3",
    "daphne>=4.2.0",
    "django>=5.2.1",
    "django-cors-headers>=4.7.0",
    "djangorestframework>=3.16.0",
    "djongo>=1.2.31",
    "langchain>=1.2.0",
    "langchain-community>=0.4.0",
    "langchain-mongodb>=0.11.0",
    "langchain-openai>=1.2.0",
    "langgraph>=1.1.0",
    "langgraph-checkpoint-mongodb>=0.3.0",
    "openai>=1.0.0",
    "pip>=25.1.1",
    "pydantic>=2.7",
    "pyjwt>=2.10.1",
    "pymongo>=4.13.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0",
    "tqdm>=4.67.1",
]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
